+++ clients/sub-account/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.0.0 - 2025-03-28
+
+### Added
+
+- `GET /sapi/v1/sub-account/futures/move-position`
+- `POST /sapi/v1/sub-account/futures/move-position`
+
 ## 1.0.0 - 2025-03-24
 
-Initial release
+- Initial release